requirements.txt
python-telegram-bot==6.1.0
requests==2.18.1
simplejson==3.11.1
jsonpickle==0.9.4
